Change how configuration management works
Configurations are now managed in a more modular way. This allows for easier
management of configurations and settings. This also allows for easier
management of hooks and settings.

This change is a breaking change. The configuration files have been moved to
the src/config directory. The configuration files have been split into
config.js, constants.js, hooks.js, hooksSettings.js, and moduleSettings.js.

The function of each file is as follows:
- config.js: This file contains the main configuration for the module. It is built
  from the constants, hooks, hooksSettings, and moduleSettings files.
- constants.js: This file contains the constants for the module. Most configurations
    should be handled here.
- hooks.js: This file contains the custom hooks used by the module, as well as the
    default hooks used by the module. It serves both as a way of ensuring consistency
    in hook names and as a way of documenting the hooks used by the module.
- hooksSettings.js: This file contains the settings for the hooks used by the module.
    It is used to configure the hooks used by the module.
- moduleSettings.js: This file contains the client-facing settings for the module.