chore: remove docker directory
The docker feature for testing has been promoted to its own repository:
https://github.com/Xain3/foundryvtt-dev-pod

Removed the docker directory to avoid duplication.

Once the new repository is stable, we can consider adding it as a git
submodule to this repository for easier access, or as an imported
package if npm supports that in the future.